+++ sentry-kotlin-multiplatform/src/appleTest/kotlin/io/sentry/kotlin/multiplatform/BaseSentryTest.kt
@@ -6,4 +6,8 @@ actual abstract class BaseSentryTest {
     actual fun sentryInit(optionsConfiguration: OptionsConfiguration) {
         Sentry.init(optionsConfiguration)
     }
+
+    actual fun sentryInitWithPlatformOptions(platformOptionsConfiguration: PlatformOptionsConfiguration) {
+        Sentry.initWithPlatformOptions(platformOptionsConfiguration)
+    }
 }