+++ sentry-kotlin-multiplatform/src/commonTest/kotlin/io/sentry/kotlin/multiplatform/PlatformOptions.kt
@@ -20,3 +20,5 @@ interface CommonPlatformOptions {
 expect interface PlatformOptions : CommonPlatformOptions
 
 expect fun createPlatformOptions(): PlatformOptions
+
+expect fun createSentryPlatformOptionsConfiguration(): PlatformOptionsConfiguration