+++ sentry-kotlin-multiplatform/src/commonTvWatchMacOsMain/kotlin/io/sentry/kotlin/multiplatform/SentryPlatformOptions.tvwatchmacos.kt
@@ -1,0 +1,8 @@
+package io.sentry.kotlin.multiplatform
+
+import io.sentry.kotlin.multiplatform.extensions.toCocoaOptionsConfiguration
+
+public actual typealias SentryPlatformOptions = cocoapods.Sentry.SentryOptions
+
+internal actual fun SentryOptions.toPlatformOptionsConfiguration(): PlatformOptionsConfiguration =
+    toCocoaOptionsConfiguration()