+++ sentry-kotlin-multiplatform/src/jvmMain/kotlin/io/sentry/kotlin/multiplatform/SentryPlatformInstance.jvm.kt
@@ -1,0 +1,9 @@
+package io.sentry.kotlin.multiplatform
+
+import io.sentry.Sentry as JvmSentry
+
+internal actual class SentryPlatformInstance : SentryInstance {
+    override fun init(configuration: PlatformOptionsConfiguration) {
+        JvmSentry.init(configuration)
+    }
+}