+++ CHANGELOG.md
@@ -2,6 +2,44 @@
 
 ## Unreleased
 
+### Features
+
+- Allow initializing the KMP SDK with native options ([#221](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/221))
+  - This allows you to initialize the SDK with platform-specific options that may not be available in the common code of the KMP SDK yet.
+
+Usage:
+```kotlin
+// build.gradle.kts
+kotlin {
+    sourceSets {
+        all {
+          languageSettings.optIn("kotlinx.cinterop.ExperimentalForeignApi")
+        }
+    }
+}
+
+// commonMain
+fun init() {
+  Sentry.initWithPlatformOptions(createPlatformOptions())
+}
+
+expect fun platformOptionsConfiguration(): PlatformOptionsConfiguration
+
+// iOS
+actual fun createPlatformOptions(): PlatformOptionsConfiguration = { 
+    dsn = "your_dsn"
+    release = "1.0.0"
+    // ...
+}
+
+// Android
+actual fun createPlatformOptions(): PlatformOptionsConfiguration = {
+  dsn = "your_dsn"
+  release = "1.0.0"
+  // ...
+}
+```
+
 ### Dependencies
 
 - Bump Java SDK from v7.8.0 to v7.9.0 ([#219](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/219))